batch/pyproject.toml
[tool.poetry]
name = "my-batch"
version = "0.1.0"
description = ""
authors = ["shiruta <hiruta@totalsolution.biz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
